+++ packages/vant/src/utils/index.ts
@@ -4,7 +4,6 @@ export * from './dom';
 export * from './create';
 export * from './format';
 export * from './constant';
-export * from './validate';
 export * from './interceptor';
 export * from './with-install';
 export * from './closest';
